no file. no cryptographic hash.

+++ sha1sum.rb
@@ -32,23 +32,28 @@ end
 
 file = "#{ARGV[0]}"
 
-# if the files is bigger than LONG_MAX then read it in 512 byte chunks
-# according to http://ruby.about.com/od/advancedruby/ss/Cryptographic-Hashes-In-Ruby.htm
-if File.size(file) > 2147483647 then
-  File.open(file) do|big_file|
-    buffer = ''
-    # read the file 512 bytes at a time
-    while not big_file.eof
-      big_file.read(512, buffer)
-      digest.update(buffer)
-    end
-  end
-  digest_2 = digest.digest
-else
-  digest_2 = digest.digest(File.read(file))
-end
+if File.exists?(file) then
 
-pp pick
-pp ARGV
-pp File.size(file)
-pp digest_2.unpack('H*').first
+  # if the files is bigger than LONG_MAX then read it in 512 byte chunks
+  # according to http://ruby.about.com/od/advancedruby/ss/Cryptographic-Hashes-In-Ruby.htm
+  if File.size(file) > 2147483647 then
+    File.open(file) do|big_file|
+      buffer = ''
+      # read the file 512 bytes at a time
+      while not big_file.eof
+        big_file.read(512, buffer)
+        digest.update(buffer)
+      end
+    end
+    digest_2 = digest.digest
+  else
+    digest_2 = digest.digest(File.read(file))
+  end
+
+  pp pick
+  pp ARGV
+  pp File.size(file)
+  pp digest_2.unpack('H*').first
+else
+  abort (message="file not found")
+end
